Popup tweaks done
