+++ frontend/src/Components/FindProduct.jsx
@@ -1,15 +1,17 @@
 import { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import { BounceLoader } from 'react-spinners'
+import getCookie from '../utilComponents/GetUserNameUtil';
 
 function FindProduct() {
     const [data, setData] = useState(null);
     const [name, setName] = useState("");
     const [loading, setLoading] = useState(null)
     const [respData, setResponseData] = useState()
-
+    const username = getCookie("user")
+    
     const fetchData = () => {
-        fetch(`https://onlinegoessilly-server.onrender.com/api/user-items/${name}`)
+        fetch(`http://localhost:8080/api/user-items/${name}/${username}`)
             .then(resp => resp.json())
             .then(result => {
                 setLoading(false)
@@ -20,7 +22,7 @@ function FindProduct() {
 
     const deleteData = async (id) => {
         try {
-            const response = await fetch(`https://onlinegoessilly-server.onrender.com/api/delete-user-items/${id}`, {
+            const response = await fetch(`http://localhost:8080/api/delete-user-items/${id}`, {
                 method: 'DELETE',
                 headers: { 'Content-Type': 'application/json' }
             })
